Work through the compiler warnings and take action as necessary.
This fixes most of the warnings, except for:
* Any collab code
* fontforge/ufo.c (skipped)
* encoding.c: Assigning const parameter AdobeStdEncoding to what may be free'd
  - No easy way to resolve this
* charview.c: 12910 - Should this be cast or converted?
  > mblist_nomm[i].ti.text_untranslated = mblist_nomm[i].ti.text;
* cvpalettes.c: Redefinitions of MID_*
* metricsview.c: selectUserChosenWordListGlyphs - Comparing GArray pointer against 0 / -1 / -2 ?
* windowmenu.c: Same question as for charview.c
  > e.g:
  > windowmenu.c:270:33: warning: assignment makes pointer from integer without a cast
      mb[i].ti.text_untranslated = mb[i].shortcut;

+++ fontforgeexe/cvpointer.c
@@ -1276,7 +1276,7 @@ return(false);
 	    bool preserveState = false;
 	    CVVisitAllControlPoints( cv, preserveState,
 	    			     FE_touchControlPoint,
-	    			     (void*)cv->b.layerheads[cv->b.drawmode]->order2 );
+	    			     (void*)(intptr_t)cv->b.layerheads[cv->b.drawmode]->order2 );
 	}
     }
     
@@ -1393,7 +1393,7 @@ static void touchControlPointsVisitor ( void* key,
 				 bool isnext,
 				 void* udata )
 {
-    SPTouchControl( sp, which, (int)udata );
+    SPTouchControl( sp, which, (int)(intptr_t)udata );
 }
 
 int CVMouseMovePointer(CharView *cv, GEvent *event) {
@@ -1530,7 +1530,7 @@ return( false );
 	// visiting all is a hammer left below in case it might be needed.
 	CVVisitAdjacentToSelectedControlPoints( cv, false,
 						touchControlPointsVisitor,
-						(void*)cv->b.layerheads[cv->b.drawmode]->order2 );
+						(void*)(intptr_t)cv->b.layerheads[cv->b.drawmode]->order2 );
 	/* CVVisitAllControlPoints( cv, false, */
 	/* 			 touchControlPointsVisitor, */
 	/* 			 (void*)cv->b.layerheads[cv->b.drawmode]->order2 ); */
